+++ components/form/BaseCheckbox.vue
@@ -1,4 +1,4 @@
-<script setup lang="ts" generic="T extends any = boolean">
+<script setup lang="ts" generic="T extends unknown = boolean">
 defineOptions({
   inheritAttrs: false,
 })
@@ -101,6 +101,10 @@ const props = withDefaults(
     classes: () => ({}),
   },
 )
+
+defineSlots<{
+  default(): any
+}>()
 
 const [modelValue] = defineModel<T | T[]>()