+++ components/form/BaseCheckboxHeadless.vue
@@ -1,4 +1,4 @@
-<script setup lang="ts" generic="T extends any = boolean">
+<script setup lang="ts" generic="T extends unknown = boolean">
 defineOptions({
   inheritAttrs: false,
 })
@@ -41,6 +41,10 @@ const props = withDefaults(
 
 const [modelValue] = defineModel<T | T[]>()
 
+defineSlots<{
+  default(props: { value: T | T[] | undefined }): any
+}>()
+
 const inputRef = ref<HTMLInputElement>()
 const id = useNinjaId(() => props.id)
 
@@ -75,7 +79,7 @@ defineExpose({
         :true-value="props.trueValue"
         :false-value="props.falseValue"
         v-bind="$attrs"
-        class="peer absolute inset-0 z-20 h-full w-full cursor-pointer opacity-0"
+        class="peer absolute inset-0 z-20 size-full cursor-pointer opacity-0"
         type="checkbox"
       />
       <slot v-bind="{ value: modelValue }"></slot>