+++ components/form/BaseRadio.vue
@@ -1,4 +1,4 @@
-<script setup lang="ts" generic="T extends any = boolean">
+<script setup lang="ts" generic="T extends unknown = boolean">
 defineOptions({
   inheritAttrs: false,
 })
@@ -73,6 +73,10 @@ const props = withDefaults(
   },
 )
 const [modelValue] = defineModel<T>()
+
+defineSlots<{
+  default(): any
+}>()
 
 const color = useNuiDefaultProperty(props, 'BaseRadio', 'color')